requirements.txt
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
Authlib==1.6.3
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.7
distro==1.9.0
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.107.0
packaging==25.0
passlib==1.7.4
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
redis==6.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
